+++ Patches/ChatMessageSystemPatch.cs
@@ -13,13 +13,11 @@ namespace Bloodcraft.Patches;
 [HarmonyPatch]
 internal static class ChatMessageSystemPatch
 {
-    static EntityManager EntityManager => Core.EntityManager;
-
     static readonly bool _eclipse = ConfigService.Eclipse;
 
     static readonly Regex _regexMAC = new(@";mac([^;]+)$");
 
-    [HarmonyBefore("gg.deca.Bloodstone")]
+    [HarmonyBefore("gg.deca.Bloodstone", "CrimsonChatFilter")]
     [HarmonyPatch(typeof(ChatMessageSystem), nameof(ChatMessageSystem.OnUpdate))]
     [HarmonyPrefix]
     static void OnUpdatePrefix(ChatMessageSystem __instance)
@@ -39,7 +37,7 @@ internal static class ChatMessageSystemPatch
 
                 if (CheckMAC(chatMessageEvent.MessageText.Value, out string originalMessage))
                 {
-                    Core.Log.LogWarning($"[ChatMessageSystem] Handling Eclipse client message...");
+                    // Core.Log.LogWarning($"[ChatMessageSystem] Handling Eclipse client message...");
                     EclipseService.HandleClientMessage(originalMessage);
                     entity.Destroy(true);
                 }